move several vars to _config.yml, add HRT startdate and info on index.html, and bump NPM & Ruby Gem packages

+++ _scripts/post_sym_ln.sh
@@ -4,5 +4,5 @@ cd ./_site/ || exit 1
 # Link
 #for d in "advguildRoot" "gameRes/TIC TAC TOE" "mc" "music.git" "torrents" "uploads" ; do
 for d in "advguildRoot" "gameRes/TIC TAC TOE" "torrents" "uploads" ; do
-    ln --symbolic --relative "../${d}" "${d}"
+    [ -f "$d" ] || [ -d "$d" ] && ln --symbolic --relative "../${d}" "${d}"
 done
